refactor(app): Convert properties to protected for trait/child access
The visibility of `$scripts`, `$inline_scripts`, and `$deferred_scripts` properties in `EnqueueAbstract.php` and the function `process_single_script` have been changed from `private` to `protected`.

This change is made to allow child classes and upcoming traits (as part of the  refactoring) to directly access and manage these arrays, improving extensibility and supporting a more modular design.